FIX: refonte de la géréniciter

+++ api/src/main/java/com/projectspring/api/Repositories/UserRepositories.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.projectspring.api.Models.UserEntities;
 
-public interface UserRepositories extends JpaRepository<UserEntities, Integer> {
+public interface UserRepositories extends JpaRepository<UserEntities, Long> {
     
     UserEntities findByUsername(String username);
 }
